requirements.txt
Flask
flask-restx
sentence-transformers
qdrant-client
requests
openai
google-generativeai
numpy
